fix: split eipfs stack
